# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "This is a sample server Arkadia Lending API. Pass Content-Type application/json\
    \ to hashed path."
  version: "1.0.0"
  title: "Arkadia Lending API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "token.su2.io"
basePath: "/arkadia"
tags:
- name: "hashed"
  description: "places hash of data on chain"
  externalDocs:
    description: "Find out more"
    url: "https://token.su2.io/arkadia"
- name: "unhashed"
  description: "places data directly on chain"
  externalDocs:
    description: "Find out more"
    url: "https://token.su2.io/arkadia"
schemes:
- "https"
- "http"
paths:
  /hashed:
    post:
      tags:
      - "hashed"
      summary: "Create hashed"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created hashed object"
        required: true
        schema:
          $ref: "#/definitions/Borrower"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Hashed"
  /unhashed/addBorrower:
    post:
      tags:
      - "unhashed"
      summary: "Creates list of users with given input array"
      description: ""
      operationId: "createUsersWithArrayInput"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "List of user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Borrower"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Unhashed"
  /unhashed/addBorrowerPayment:
    post:
      tags:
      - "unhashed"
      summary: "Creates list of users with given input array"
      description: ""
      operationId: "addBorrowerPayment"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "List of user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Payment"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Unhashed"
  /unhashed/addBorrowerLoan:
    post:
      tags:
      - "unhashed"
      summary: "Creates list of users with given input array"
      description: ""
      operationId: "addBorrowerLoan"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "List of user object"
        required: true
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Loan"
      responses:
        default:
          description: "successful operation"
      x-swagger-router-controller: "Unhashed"
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Borrower:
    type: "object"
    properties:
      borrower:
        type: "integer"
        format: "int64"
        description: "Borrower Id"
      industry:
        type: "string"
        description: "Example: Steel"
      countryCode:
        type: "string"
        description: "Three digits: Example: USA"
      assets:
        type: "integer"
        format: "int64"
        description: "Example: 100"
      debt:
        type: "integer"
        format: "int64"
      profit:
        type: "integer"
        format: "int64"
      expectedCashFlow:
        type: "integer"
        format: "int32"
    example:
      assets: 6
      expectedCashFlow: 5
      countryCode: "countryCode"
      borrower: 0
      industry: "industry"
      debt: 1
      profit: 5
    xml:
      name: "Borrower"
  Payment:
    type: "object"
    properties:
      borrower:
        type: "integer"
        format: "int64"
        description: "Borrower Id"
      industry:
        type: "string"
        description: "Example: Steel"
      countryCode:
        type: "string"
        description: "Three digits: Example: USA"
      assets:
        type: "integer"
        format: "int64"
        description: "Example: 100"
      debt:
        type: "integer"
        format: "int64"
      profit:
        type: "integer"
        format: "int64"
      expectedCashFlow:
        type: "integer"
        format: "int32"
    xml:
      name: "Payment"
  Loan:
    type: "object"
    properties:
      borrower:
        type: "integer"
        format: "int64"
        description: "Borrower Id"
      industry:
        type: "string"
        description: "Example: Steel"
      countryCode:
        type: "string"
        description: "Three digits: Example: USA"
      assets:
        type: "integer"
        format: "int64"
        description: "Example: 100"
      debt:
        type: "integer"
        format: "int64"
      profit:
        type: "integer"
        format: "int64"
      expectedCashFlow:
        type: "integer"
        format: "int32"
    xml:
      name: "Loan"
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      type:
        type: "string"
      message:
        type: "string"
externalDocs:
  description: "For more detailed information"
  url: "https://github.com/arkadialend/Proof-Of-Concept"